Remove dead code in end_psymtab
I noticed that there is a bit of dead code in end_psymtab.
This deletes it.

Normally I would investigate a fix for the code.  However, considering
that the code has been this way a long time (since the first import to
sourceware) and considering that dbxread.c is not as important any
more, I think it's safe to just consider that there's no bug.

gdb/ChangeLog
2018-06-07  Tom Tromey  <tromey@redhat.com>

	* dbxread.c (end_psymtab): Remove dead code.