Read pseudo registers from frame instead of regcache
Reading pseudo registers (made of one or more raw register) from an
upper stack frame does not work.  The raw registers needed to compose
the pseudo registers are always read from the current thread's regcache,
which is effectively frame 0.

Here's an example using the x86-64 test included in this patch (with
some output remove for brevity):

  (gdb) tb break_here_asm
  (gdb) r
  (gdb) info registers rbx ebx
  rbx            0x2021222324252627  2315169217770759719
  ebx            0x24252627          606414375
  (gdb) up
  (gdb) info registers rbx ebx
  rbx            0x1011121314151617  1157726452361532951
  ebx            0x24252627          606414375

We would expect the last ebx value to be 0x14151617, half of the rbx value in
the same frame.

The problem comes from the fact that the gdbarch hooks:

  - pseudo_register_read
  - pseudo_register_read_value
  - pseudo_register_write

receive a regcache as a parameter, as a way to get the values of the raw
registers required to compute the pseudo register value.  However, the
regcache object always contains the current register values, not the
values unwound to the frame we're interested in.  That's why ebx in
frame #1 appears to have the value it has in frame #0.

This patch introduces new implementations of the register_reader and
register_readwriter interfaces which access register values by unwinding
them from a particular frame.  It therefore allows computing pseudo
register values using the raw register values of that frame.

My initial implementation made it so raw registers were unwound using
the frame unwinder (generally goes to dwarf2_frame_prev_register) and
pseudo registers were unwound through gdbarch_pseudo_register_read.  The
problem with this, as pointed out by Ulrich, is that we could have DWARF
unwind info for pseudo registers sometimes.

An example of this are the S registers on ARM, where two 32-bits S
registers form one 64-bits D register.  According to the "DWARF for the
ARM architecture" document, producers should describe the location of
the saved D registers and the value of the S registers should be
inferred from them.  In GDB words, S are pseudo registers build from D,
which are raw registers.  So we would unwind D using the frame unwinder
(i.e. DWARF info) and unwind S using gdbarch_pseudo_register_read, which
will actually read the right D register and derive the S value from
that.

However, before the introduction of the D registers, the S registers had
register numbers assigned (64-95).  So it's possible to find binaries
whose DWARF info have rules to unwind S registers.  Using the naive
approach described previously wouldn't work, because we would try to
read the matching D register, for which we don't have any unwind
information, and end up with a wrong result.

So what this patch does is first try to unwind pseudo registers using
the frame unwinder.  If we specifically have info for that register, we
use that.  Otherwise, we fall back to the common case of going through
gdbarch_pseudo_register_read.

This change affected the cooked_write_test and cooked_read_test tests.
I had to set up a bit more fake global state to get them to pass.  This
is because reading/writing pseudo registers now involve frames, and in
particular the sentinel frame, which was not created before.