Fix some test build failures with clang in gdb.python
Fix this when running py-finish-breakpoint.exp

.../py-finish-breakpoint.c:59:1: warning: control reaches end of non-void function
      [-Wreturn-type]
}
^

and these when running py-pp-integral.exp/py-pp-re-notag.exp

.../py-pp-integral.c:23:3: warning: expression result unused [-Wunused-value]
  *t++;
  ^~~~

.../py-pp-re-notag.c:23:3: warning: expression result unused [-Wunused-value]
  *t++;
  ^~~~

I think the changes in this patch don't affect what the tests are
exercising.  In the second case, I believe that the code in this patch
(with the parentheses) is what was initially intended, although it
doesn't really influence the tests.

gdb/ChangeLog:

		* gdb.python/py-finish-breakpoint.c (call_longjmp_1,
		call_longjmp): Return void.
		* gdb.python/py-pp-integral.c (tick_tock): Add
		parentheses.
		* gdb.python/py-pp-re-notag.c (tick_tock): Likewise.