Compile gdb.python c++ test cases using a .cpp extension
Compiling .c files as C++ makes clang unhappy:

  warning: treating 'c' input as 'c++' when in C++ mode, this behavior is deprecated [-Wdeprecated]

One way to inhibit this warning is to pass -xc++ to the compiler (as we
do for the GDB source code).  However, I did not find an easy way to do
this.  Passing CXX_FOR_TARGET="clang++ -xc++" does not work, because
that is used for linking too, and -xc++ should not be passed when
linking.  Using additional_flags=-xc++ does not work, because these
flags are passed to the compiler after the source file, but to have an
effect, it should be placed before.

So this patch makes it so when we want to compile a C file as C++
(usually when the same file is used for both C and C++ tests), we copy
the file to have a .cpp extension and compile that.

gdb/testsuite/ChangeLog:

	* gdb.python/py-mi.exp: Compile C++ file using .cpp extension.
	* gdb.python/py-prettyprint.exp: Likewise.
	* gdb.python/py-symbol.exp: Likewise.
	* gdb.python/py-type.exp: Likewise.
	* gdb.python/py-value.exp: Likewise.