Use -qualified flag when using the start command
When using the "start" command, GDB puts a temporary breakpoint on the
"main" symbol (we literally invoke the tbreak command).  However, since
we do wild matching by default, it puts a breakpoint on any C++ method
called main.  For example, when debugging GDB, it creates a total of 24
locations:

  (gdb) start
  Temporary breakpoint 1 at 0x198c1e9: main. (24 locations)

as there are a bunch of methods called main in the selftests, such as

  selftests::string_view::capacity_1::main()

The dummiest, most straightforward solution is to add -qualified when
invoking tbreak.  With this patch, "start" creates a single-location
breakpoint, as expected.

	* infcmd.c (run_command_1): Pass -qualified to tbreak when usind
	the "start" command.