Update gnulib import
This patch updates our gnulib import to the latest gnulib commit.  The
motivation behind this is that I'll propose (in a separate thread) to
use the vcs_to_changelog.py script in GDB, which is distributed as a
gnulib module.

Having an up to date gnulib will help getting the latest version of that
script.  I also think that this is a good time to update gnulib, since
we are more of less at the beginning of the GDB 10 development cycle, so
we have time to iron out bugs that this could introduce.

I have tried to send it on the buildbot, but I haven't received a
response, so I presume it's because the patch is too big.

This patch is also available on this branch:

  https://github.com/simark/binutils-gdb/tree/update-gnulib-import

gnulib/ChangeLog:

	* update-gnulib.sh (GNULIB_COMMIT_SHA1): Update.
	* import/*: Re-generate.
	* Makefile.in: Re-generate.
	* aclocal.m4: Re-generate.
	* config.in: Re-generate.
	* configure: Re-generate.