gdb: Move construct_inferior_arguments to gdbsupport
This moves the function construct_inferior_arguments from
gdb/inferior.h and gdb/infcmd.c to gdbsupport/common-inferior.{h,cc}.
While at it, also move the function's comment to the header file
to align with current standards.

The intention is to use it from gdbserver in a follow-up commit.

gdb/ChangeLog:

2020-05-12  Michael Weghorn  <m.weghorn@posteo.de>

        * infcmd.c, inferior.h: (construct_inferior_arguments):
          Moved function from here to gdbsupport/common-inferior.{h,cc}

gdbsupport/ChangeLog:

2020-05-12  Michael Weghorn  <m.weghorn@posteo.de>

        * common-inferior.h, common-inferior.cc: (construct_inferior_arguments):
        Move function here from gdb/infcmd.c, gdb/inferior.h