Handle multiple target events before commit resume
Hi all,

Here's a small patch to try, following our discussion the other day about
grouping resumptions requests more.  This is the smallest change I can
think of that does what we want.

It is not completely correct, there are regressions in the testsuite when running
on Linux native (and it is a bit of a naive change).  But it works well enough for
standard cases using the remote target.  I tried it with the AMD GPU target as well,
and according to the logs its does what it's supposed to do.  Hopefully, it works well
enough for you to try it and see how much it impacts performance.

We can then discuss if we want to go further down that path or find other approaches.

Simon

From 26b4fcb83d6b63926828c1cb686b38eb6d502942 Mon Sep 17 00:00:00 2001
From: Simon Marchi <simon.marchi@efficios.com>
Date: Thu, 7 May 2020 11:48:31 -0400
Subject: [PATCH] gdb: commit resumes after having consumed all events

Modify fetch_inferior_event to fetch (and handle) events from the target
as long as it provides events.

When resuming a thread that has hit a breakpoint (and it was decided
that it should be resumed), don't commit_resume just yet.  Wait until
the target has no more events to provide to do so.

Change-Id: Ia223a7d95dd5a4844c6a96e1125fb03d3e295584