gdb/jit: use objfile_key instead of objfile_data
Replace the use of objfile_data with the newer objfile_key.  The
advantage is that the methods are type-safe, so we are likely to
introduce a bug by casting to the wrong type.

The free_objfile_data function naturally becomes the destructor of
jit_objfile_data.  However, free_objfile_data accesses the objfile to
which the data is attached, which the destructor of jit_objfile_data
doesn't have access to.  To work around this, I added a backlink to the
owning objfile in jit_objfile_data.  This is however temporary, it goes
away in a subsequent patch.

gdb/ChangeLog:

	* jit.c (jit_objfile_data): Remove.
	(jit_objfile_data) <jit_objfile_data>: New constructor.
	<~jit_objfile_data>: New destructor, code moved from
	free_objfile_data.
	<objfile>: New field.
	(jit_per_objfile): New.

Change-Id: I66822dc10315a189c8d108f03fdb2478764819bd