gdb/jit: split jit_objfile_data in two
The jit_objfile_data is currently used to hold information about both
objfiles that are the result of JIT compilation (JITed) and objfiles
that can produce JITed objfiles (JITers).  I think that this double use
of the type is confusing, and that things would be more obvious if we
had one type for each role.

This patch splits it into:

- jited_objfile_data: for data about an objfile that is the result of a
  JIT compilation
- jiter_objfile_data: for data about an objfile which produces JITed
  objfiles

There are now two per-objfiles registries, one for each kind.  With this
change, the following invariants hold:

- an objfile has an entry in the registry `jit_per_jiter_objfile` iff it
  defines the required symbols of the JIT interface
- an objfile has an entry in `jit_per_jited_objfile` iff it is the
  product of JIT compilation (has been produced by some JITer)

gdb/ChangeLog:

	* jit.c (jit_objfile_data): Split into...
	(jiter_objfile_data): ... this ...
	(jited_objfile_data): ... and this.
	(jit_per_objfile): Remove.
	(jit_per_jiter_objfile, jit_per_jited_objfile): New.
	(get_jit_objfile_data): Rename to...
	(get_jiter_objfile_data): ... this.
	(jit_read_descriptor): Use get_jiter_objfile_data.
	(jit_find_objf_with_entry_addr): Use jit_per_jited_objfile.
	(jit_breakpoint_re_set_internal): Use get_jiter_objfile_data.
	(jit_inferior_exit_hook): Use jit_per_jited_objfile.

Change-Id: I1c5fe863aa090bf8e401b6814461860efbbaec39